add custom hook to clean the code
