+++ frontend/src/components/pages/PageTripsLists.tsx
@@ -1,27 +1,10 @@
-import {useEffect, useState} from "react";
-import axios from "axios";
-import {Trip} from "../../model/Trip.ts";
 import TripsList from "../TripsList.tsx";
 import {getTimeGroupedTrips} from "../../util/getTimeGroupedTrips.ts";
+import {useFetchTrips} from "../../hooks/useFetchTrips.ts";
 
 export default function PageTripsLists() {
-    const [trips, setTrips] = useState<Trip[] | null>(null)
 
-    useEffect(() => {
-        async function fetchTrips() {
-            try {
-                const response = await axios.get("/api/trips")
-                if (response.status === 200) {
-                    const tripsData = await response.data
-                    setTrips(tripsData)
-                }
-            } catch (err) {
-                console.error(err)
-            }
-        }
-
-        fetchTrips()
-    }, [])
+    const [trips] = useFetchTrips()
 
 
     if (!trips) {
@@ -30,11 +13,11 @@ export default function PageTripsLists() {
         )
     }
 
-    const {pastTrips, onGoingTrips, futureTrips} = getTimeGroupedTrips(trips)
+    const {pastTrips, onGoingTrip, futureTrips} = getTimeGroupedTrips(trips)
 
     return (
         <>
-            <TripsList title="On going Trip" list={onGoingTrips} />
+            <TripsList title="On going Trip" list={onGoingTrip} />
             <TripsList title="Future Trips" list={futureTrips} />
             <TripsList title="Past Trips" list={pastTrips} />
         </>
